Fixed small bugs

+++ group3_sbs/external/views.py
@@ -861,7 +861,6 @@ def rejectAccessRequests(request):
 @user_passes_test(is_external_user)
 def reject_approvals(request):
     user = request.user
-    #add this to transactions of the merchant as failed ones
     string_transaction_id = str(request.POST['id'])
     transaction_id = int(string_transaction_id)
     transaction = MerchantPaymentRequest.objects.all().filter(id=transaction_id)
@@ -936,29 +935,13 @@ def showPaymentRequests(request):
 @user_passes_test(is_external_user)
 def update_approvals(request):
     user = request.user
-
     error_redirect = 'external:error'
-    success_redirect = 'external:showPaymentRequests' #Show transaction requests
-
-    #add this to transactions of the user --not implemented
-
-    # Remove this transaction from requests
+    success_redirect = 'external:showPaymentRequests'
     string_transaction_id = str(request.POST['id'])
     transaction_id = int(string_transaction_id)
     transaction = MerchantPaymentRequest.objects.all().filter(id=transaction_id).first()
-
-    #sender account type
-    account_type = str(transaction.accountType)
-
-#    if not payment_merchant_request_validate(request=request, merchant_request=transaction):
-#        return HttpResponseRedirect(reverse('external:error'))
+    transaction = MerchantPaymentRequest.objects.all().filter(id=transaction_id).first()
+    if not payment_merchant_request_validate(request=request, merchant_request=transaction):
+        return HttpResponseRedirect(reverse(error_redirect))
     transaction.delete()
-    checkingRequests = MerchantPaymentRequest.objects.all().filter(accountType="Checking").filter(
-        clientAccountNum=user.individualcustomer.checking_account_id)
-    savingRequests = MerchantPaymentRequest.objects.all().filter(accountType="Savings").filter(
-        clientAccountNum=user.individualcustomer.savings_account_id)
-
-
-    #make Transfer
-    type_of_transaction = TRANSACTION_TYPE_TRANSFER
-    return transfer_validate(request=request, type_of_transaction=type_of_transaction, account_type=account_type, success_redirect=success_redirect, error_redirect=error_redirect)
+    return HttpResponseRedirect(reverse(success_redirect))
